This script goes through and discovers potential car bomb events on twitter, validates and then posts them to twitter.com/weaware.
The goal is to use this to connect a community who cares about a cause
by creating interactions with realtime news they care about. Two ideas:

1) A wristband that heats up whenever there is something posted to
@weaware
2) A mesopotamian artifact placed between two clamps that gets
tightened a half turn every time there’s a post to @weaware
3) A mobile chatroom app that opens a notification for each validated
event that allows for conversation to lead to action

I need a lot of help to flesh this out. I’m not sure how to put this on
a server and have it run continuously without using a chron job. Also
the twitter API is limited and I(we) may need to write our own news
scraper, in house validator, push notification thing.

There are some questions in the code that someone with better
python/twitter API skills can help with
1) How do we use the advanced search functionality in twython?
2) Rate limiting questions. Is this going to be impossible to do on
twitter? Can someone with connections to twitter help me request a
higher rate?